refactor(http): clean code - split controllers / middleware (#90)
* chore(http): use http-server folder

* chore(http): move root route

* test(http): add first test

* chore: add updated package.json

* chore(http): finish root tests

* chore: create context (AsyncLocalStorage)

* feat(http): split data endpoint

* chore(http): port middleware in split file

* chore(http): remove useless spec test

* chore(http): fix naming for root import

* chore(http): add flags

* revert: remove sinon deps

* chore: add missing import comment
