+++ src/http-server/context.js
@@ -1,0 +1,3 @@
+import { AsyncLocalStorage } from "async_hooks";
+
+export const context = new AsyncLocalStorage();
